+++ Printed_Digit_Training/load_datasets.py
@@ -1,14 +1,12 @@
 import numpy as np
 import cv2
 import os
+from sklearn.model_selection import train_test_split
 
 
 def load_printed_digit_dataset():
-    path = 'Datasets/Digits'
-    imageDimensions = (32, 32, 3)
-    batchSizeVal = 50
-    epochsVal = 2
-    stepsPerEpochVal = 2000
+    path = '../Datasets/Digits'
+    train_size, test_size = 0.80, 0.20
 
     all_images = []
     class_number = []
@@ -36,5 +34,9 @@ def load_printed_digit_dataset():
     all_images = np.array(all_images)
     class_number = np.array(class_number)
 
-    return all_images, class_number
+    train_images, test_images, train_labels, test_labels = train_test_split(all_images, class_number,
+                                                                            train_size=train_size, test_size=test_size)
 
+    del all_images, class_number
+
+    return train_images, test_images, train_labels, test_labels, number_of_classes
